Tidied up for static analysis

+++ src/Math/Deviation.cpp
@@ -21,10 +21,10 @@ void Deviation::Set(float sumOfSquares, float sum, size_t numPoints) noexcept
 	}
 	else
 	{
-		mean = sum/numPoints;
-		const float square = sumOfSquares/numPoints - fsquare(mean);
+		mean = sum/(float)numPoints;
+		const float square = sumOfSquares/(float)numPoints - fsquare(mean);
 		// 'square' can be slightly less than 0 due to rounding error
-		deviationFromMean = (square <= 0) ? 0.0 : fastSqrtf(square);
+		deviationFromMean = (square <= 0.0) ? 0.0 : fastSqrtf(square);
 	}
 }
